requirements.txt
PyQt6>=6.0.0
PyYAML>=6.0 # For reading config.yaml
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
ffmpeg-python>=0.2.0
openai>=1.0.0
protobuf>=3.19.6 # Pin to version < 3.20 compatible with TF/Tensorboard & >= 3.19 for generated code
pyannote.audio>=3.1.1 # Use specific version aligned with model
pydub>=0.25.1
pytest>=7.4.0
pytest-mock>=3.11.1
python-dotenv>=1.0.0
requests>=2.31.0
spacy>=3.7.0
speechbrain>=1.0.0
torch>=2.0.0 # Check compatibility with pyannote.audio version
torchaudio>=2.0.0 # Check compatibility with torch
tqdm>=4.66.0
yt-dlp>=2025.3.31
Django>=4.2
djangorestframework>=3.14
